Routine commit - Introduce ServiceStarterDialog - Introduce MaterialButtonSpreadContainer - Add place holder view to FlowEditorDialog - Remove useless files - Improve UI/UX

+++ app/src/main/java/top/xjunz/tasker/ktx/Color.kt
@@ -6,6 +6,6 @@ import androidx.core.graphics.ColorUtils
 /**
  * @author xjunz 2022/12/11
  */
-fun Int.alphaModified(@FloatRange(from = 0.0, to = 1.0) alpha: Double): Int {
+fun Int.alphaModified(@FloatRange(from = 0.0, to = 1.0) alpha: Float): Int {
     return ColorUtils.setAlphaComponent(this, (alpha * 0xFF).toInt())
 }
